% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_ls.R
\name{gh_branches}
\alias{gh_branches}
\title{Get the name of all branches of a GitHub repository.}
\usage{
gh_branches(path)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir_1/subdir_2/subdir_n".}
}
\value{
A character string.
}
\description{
Get the name of all branches of a GitHub repository.
}
\examples{
gh_branches("r-lib/usethis")
}
\seealso{
ghr_fields

Other functions to get github responses: \code{\link{gh_get}}
}
\concept{functions to get github responses}
