% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{gh_expand_home}
\alias{gh_expand_home}
\title{Expand GitHub paths.}
\usage{
gh_expand_home(path, branch = "master")
}
\arguments{
\item{branch}{Character vector giving the branch name of a GitHub repo.}
}
\value{
Character vector.
}
\description{
Expand GitHub paths.
}
\examples{
gh_expand_home("tidyverse")
gh_expand_home("tidyverse/dplyr")
gh_expand_home("tidyverse/dplyr/R")
}
