% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_ls.R
\name{gh_get}
\alias{gh_get}
\title{Get a response from the GitHub API.}
\usage{
gh_get(path, ref = "master")
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir_1/subdir_2/subdir_n".}

\item{ref}{Branch name.}
}
\value{
A list.
}
\description{
Get a response from the GitHub API.
}
\examples{
gh_get("r-lib/usethis")

# The request to GitHub happens only the first time you call gh_get()
system.time(gh_get("r-lib/usethis/R"))
# Later calls take no time because the first call is memoised
system.time(gh_get("r-lib/usethis/R"))

ghr_path(gh_get("r-lib/usethis", ref = "gh-pages"))
# Same
ghr_path(gh_get("r-lib/usethis@gh-pages"))
ghr_path(gh_get("r-lib/usethis/news@gh-pages"))
}
\seealso{
ghr_fields

Other functions to get github responses: \code{\link{gh_branches}}
}
\concept{functions to get github responses}
