% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_ls.R
\name{gh_ls}
\alias{gh_ls}
\title{List files on GitHub and their html URLs.}
\usage{
gh_ls(path)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir_1/subdir_2/subdir_n".}
}
\value{
* `gh_ls()` returns a character vector giving the in a GitHub directory.
  * `gh_html()` returns the html URLs of the files in a GitHub directory.
}
\description{
List files on GitHub and their html URLs.
}
\examples{
gh_ls("r-lib")
gh_ls("r-lib/gh")
gh_ls("r-lib/gh/tests")
gh_ls("r-lib/gh/tests/testthat")

gh_html("r-lib")
gh_html("r-lib/gh")
gh_html("r-lib/gh/tests")
gh_html("r-lib/gh/tests/testthat")
}
