% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_ls.R
\name{gh_path}
\alias{gh_path}
\title{Create a request for GitHub contents that gh::gh() understands.}
\usage{
gh_path(path)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir_1/subdir_2/subdir_n".}
}
\value{
A character vector.
}
\description{
Create a request for GitHub contents that gh::gh() understands.
}
\examples{
gh_path("owner")
gh_path("owner/repo")
gh_path("owner/repo/subdir")
gh_path("owner/repo/subdir_1/subdir_2/subdir_3")
gh_path("O/R/S1/S2/S3")
}
