% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_get.R
\name{ghr_get}
\alias{ghr_get}
\title{Get a response from the GitHub API.}
\usage{
ghr_get(path, ref = "master", ...)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir" or
"owner/repo/subdir@branch", e.g.: "maurolepore/ghr/reference@gh-pages".}

\item{ref}{Branch name.}

\item{...}{Arguments passed to \code{\link[gh:gh]{gh::gh()}}.}
}
\value{
A list.
}
\description{
Get a response from the GitHub API.
}
\examples{
# The request to GitHub happens only the first time you call ghr_get()
system.time(ghr_get("maurolepore/ghr"))
# Later calls take no time because the first call is memoised
system.time(ghr_get("maurolepore/ghr"))

gh_response <- ghr_get("maurolepore/ghr")
class(gh_response)
length(gh_response)
str(gh_response[[1]])

ghr_pull(ghr_get("maurolepore/ghr", ref = "gh-pages"), "path")
# Same
ghr_pull(ghr_get("maurolepore/ghr@gh-pages"), "path")
}
\seealso{
ghr_fields

Other functions to get github responses: \code{\link{ghr_show_branches}}
}
\concept{functions to get github responses}
