% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_ls.R
\name{ghr_ls}
\alias{ghr_ls}
\alias{ghr_ls_html_url}
\alias{ghr_ls_download_url}
\title{List directories, html-URLs and download-URLs in a GitHub directory.}
\usage{
ghr_ls(path, ..., regexp = NULL, ignore.case = FALSE, invert = FALSE)

ghr_ls_html_url(path, ..., regexp = NULL, ignore.case = FALSE,
  invert = FALSE)

ghr_ls_download_url(path, ..., regexp = NULL, ignore.case = FALSE,
  invert = FALSE)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir" or
"owner/repo/subdir@branch", e.g.: "maurolepore/ghr/reference@gh-pages".}

\item{...}{Arguments passed to \code{\link[gh:gh]{gh::gh()}} via \code{\link[=ghr_get]{ghr_get()}}.}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to grep() to
filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}
}
\value{
A character string.
}
\description{
List directories, html-URLs and download-URLs in a GitHub directory.
}
\examples{
path <- "maurolepore/tor/inst/extdata/mixed"

# The first call make the request
system.time(ghr_ls(path))
# Takes no time because the first call is memoised
system.time(ghr_ls(path))

ghr_ls(path, regexp = "[.]csv$")
ghr_ls(path, regexp = "[.]csv$", invert = TRUE)

ghr_ls(path, regexp = "[.]RDATA$", invert = TRUE, ignore.case = FALSE)
ghr_ls(path, regexp = "[.]RDATA$", invert = TRUE, ignore.case = TRUE)

# ghr_ls_download_url() and ghr_ls_html_url are similar
(d_url <- ghr_ls_download_url(path, regexp = "[.]csv$")[[1]])
read.csv(d_url)

(h_url <- ghr_ls_html_url(path, regexp = "[.]csv$")[[1]])
if (interactive()) {
  utils::browseURL(h_url)
}
}
\concept{shortcuts for common tasks}
