% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_ls_path.R
\name{ghr_ls_path}
\alias{ghr_ls_path}
\title{List directories, html-URLs and download-URLs in a GitHub directory.}
\usage{
ghr_ls_path(path, regexp = NULL, ignore.case = FALSE, invert = FALSE,
  recursive = FALSE)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir" or
"owner/repo/subdir@branch", e.g.: "maurolepore/ghr/reference@gh-pages".}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to grep() to
filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}

\item{recursive}{Should directories be listed recursively?
the filenames.}
}
\value{
A character string.
}
\description{
List directories, html-URLs and download-URLs in a GitHub directory.
}
\examples{
ghr_ls_path("maurolepore/ghr/tests")
ghr_ls_path("maurolepore/ghr/tests", recursive = TRUE)
}
