% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_ls_path.R
\name{ghr_ls_path_impl}
\alias{ghr_ls_path_impl}
\title{List paths in a GitHub directory.}
\usage{
ghr_ls_path_impl(path)
}
\arguments{
\item{path}{}
}
\description{
List paths in a GitHub directory.
}
\examples{
ghr_ls_path("maurolepore")
ghr_ls_path("maurolepore/tor")
ghr_ls_path("maurolepore/tor/tests")
ghr_ls_path("maurolepore/tor/tests/testthat")
}
