% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_get.R
\name{ghr_show_branches}
\alias{ghr_show_branches}
\title{Get the name of all branches of a GitHub repository.}
\usage{
ghr_show_branches(path)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir" or
"owner/repo/subdir@branch", e.g.: "maurolepore/ghr/reference@gh-pages".}
}
\value{
A character string.
}
\description{
Get the name of all branches of a GitHub repository.
}
\examples{
ghr_show_branches("maurolepore/ghr")
}
\seealso{
ghr_fields

Other functions to get github responses: \code{\link{ghr_get}}
}
\concept{functions to get github responses}
