% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_show_fields.R
\name{ghr_show_fields}
\alias{ghr_show_fields}
\alias{ghr_pull}
\title{Show and pull available fields of a GitHub response.}
\usage{
ghr_show_fields(gh_response)

ghr_pull(gh_response, field)
}
\arguments{
\item{gh_response}{A list object of class "gh_response".}

\item{field}{Character string giving the name of the the element to
extract from the github-response list.}
}
\value{
A character string.
}
\description{
Show and pull available fields of a GitHub response.
}
\examples{
gh_response <- ghr_get("maurolepore/ghr/R")

ghr_show_fields(gh_response)

ghr_pull(gh_response, "name")

# Working with non-default branches
ghr_pull(ghr_get("maurolepore/ghr", ref = "gh-pages"), "path")
# Same
ghr_pull(ghr_get("maurolepore/ghr@gh-pages"), "path")

ghr_pull(ghr_get("maurolepore/ghr/reference@gh-pages"), "path")
}
\seealso{
ghr_get gh_branches
}
\concept{functions to work with github responses}
