% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_readme.R
\name{html_file}
\alias{html_file}
\title{ghr_browse_file(ghr_get("maurolepore/ghr"), "README.md")}
\usage{
html_file(response, file)
}
\description{
ghr_browse_file(ghr_get("maurolepore/ghr"), "README.md")
}
