% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_ls.R
\name{gh_ls}
\alias{gh_ls}
\title{List files on GitHub.}
\usage{
gh_ls(path)
}
\arguments{
\item{x}{}
}
\description{
This function works on GitHub as [fs::dir_ls()] on a the file system.
}
\examples{
gh_ls("forestgeo")
gh_ls("forestgeo/fgeo")
gh_ls("forestgeo/fgeo/tests")
gh_ls("forestgeo/fgeo/tests/testhat")
gh_ls("forestgeo/fgeo/tests/bad-dir")
}
