% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_ls.R
\name{gh_path}
\alias{gh_path}
\title{Convert a path such as owner/repo/subdir into an `endpoint` for `gh::gh()`.}
\usage{
gh_path(path)
}
\arguments{
\item{path}{}
}
\value{
A character string.
}
\description{
Convert a path such as owner/repo/subdir into an `endpoint` for `gh::gh()`.
}
\examples{
gh_path("r-lib")
gh_path("r-lib/gh")
gh_path("r-lib/gh/tests")
gh_path("r-lib/gh/tests/testthat")
}
