% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr-package.R
\docType{package}
\name{ghr-package}
\alias{ghr}
\alias{ghr-package}
\title{ghr: Explore and Access GitHub Directories}
\description{
Explore and access GitHub directories using an
    intuitive syntax of the form "owner/repository/subdirectory@branch".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurolepore/ghr}
  \item Report bugs at \url{https://github.com/maurolepore/ghr/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (0000-0002-1986-7988)

}
\keyword{internal}
