% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_get.R
\name{ghr_get}
\alias{ghr_get}
\title{Get a response from the GitHub API.}
\usage{
ghr_get(path, ref = "master")
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir_1/subdir_2/subdir_n".}

\item{ref}{Branch name.}
}
\value{
A list.
}
\description{
Get a response from the GitHub API.
}
\examples{
ghr_get("r-lib/usethis")

# The request to GitHub happens only the first time you call ghr_get()
system.time(ghr_get("r-lib/usethis/R"))
# Later calls take no time because the first call is memoised
system.time(ghr_get("r-lib/usethis/R"))

ghr_pull(ghr_get("r-lib/usethis", ref = "gh-pages"), "path")
# Same
ghr_pull(ghr_get("r-lib/usethis@gh-pages"), "path")

ghr_pull(ghr_get("r-lib/usethis/news@gh-pages"), "path")
}
\seealso{
ghr_fields

Other functions to get github responses: \code{\link{ghr_show_branches}}
}
\concept{functions to get github responses}
