% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_ls.R
\name{ghr_ls}
\alias{ghr_ls}
\title{List files in a GitHub 'directory'.}
\usage{
ghr_ls(path, regexp = NULL, ignore.case = FALSE, invert = FALSE, ...)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir" or
"owner/repo/subdir@branch", e.g.: "maurolepore/ghr/reference@gh-pages".}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to grep() to
filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}

\item{...}{Arguments passed to \code{\link[gh:gh]{gh::gh()}} via \code{\link[=ghr_get]{ghr_get()}}.}
}
\value{
A character string.
}
\description{
List files in a GitHub 'directory'.
}
\examples{
path <- "maurolepore/tor/inst/extdata/mixed"

# The first call make the request
system.time(ghr_ls(path))
# Takes no time because the first call is memoised
system.time(ghr_ls(path))

ghr_ls(path, regexp = "[.]csv$")
ghr_ls(path, regexp = "[.]csv$", invert = TRUE)

ghr_ls(path, regexp = "[.]RDATA$", invert = TRUE, ignore.case = FALSE)
ghr_ls(path, regexp = "[.]RDATA$", invert = TRUE, ignore.case = TRUE)
}
