% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghr_ls.R
\name{ghr_ls}
\alias{ghr_ls}
\title{List files in a GitHub 'directory'.}
\usage{
ghr_ls(path, regexp = NULL, ignore.case = FALSE, invert = FALSE)
}
\arguments{
\item{path}{A string formatted as "owner/repo/subdir" or
"owner/repo/subdir@branch", e.g.: "maurolepore/ghr/reference@gh-pages".}

\item{regexp}{A regular expression (e.g. \link{.}csv$) passed on to grep() to
filter paths.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}
}
\value{
A character string.
}
\description{
List files in a GitHub 'directory'.
}
\examples{
# The first call make the request
system.time(ghr_ls("maurolepore/ghr/R"))
# Takes no time because the first call is memoised
system.time(ghr_ls("maurolepore/ghr/R"))

ghr_ls("maurolepore/ghr/R")
ghr_ls("maurolepore/ghr/R", regexp = "get")
ghr_ls("maurolepore/ghr/R", regexp = "get", invert = TRUE)
ghr_ls("maurolepore/ghr/R", regexp = "GET")
ghr_ls("maurolepore/ghr/R", regexp = "GET", ignore.case = TRUE)
}
